+++ packages/dts-plugin/src/core/lib/DTSManager.ts
@@ -1,7 +1,7 @@
-import ansiColors from 'ansi-colors';
 import path from 'path';
 import { rm } from 'fs/promises';
 import fs from 'fs';
+import fse from 'fs-extra';
 import {
   MANIFEST_EXT,
   Manifest,
@@ -103,21 +103,35 @@ class DTSManager {
 
     const mfTypesPath = retrieveMfTypesPath(tsConfig, remoteOptions);
 
-    if (hasRemotes) {
-      const tempHostOptions = {
-        moduleFederationConfig: remoteOptions.moduleFederationConfig,
-        typesFolder: path.join(mfTypesPath, 'node_modules'),
-        remoteTypesFolder:
-          remoteOptions?.hostRemoteTypesFolder || remoteOptions.typesFolder,
-        deleteTypesFolder: true,
-        context: remoteOptions.context,
-        implementation: remoteOptions.implementation,
-        abortOnError: false,
-      };
-      await this.consumeArchiveTypes(tempHostOptions);
+    if (hasRemotes && this.options.host) {
+      try {
+        const { hostOptions } = retrieveHostConfig(this.options.host);
+        const remoteTypesFolder = path.resolve(
+          hostOptions.context,
+          hostOptions.typesFolder,
+        );
+
+        const targetDir = path.join(mfTypesPath, 'node_modules');
+        if (fs.existsSync(remoteTypesFolder)) {
+          const targetFolder = path.resolve(remoteOptions.context, targetDir);
+          await fse.ensureDir(targetFolder);
+          await fse.copy(remoteTypesFolder, targetFolder, { overwrite: true });
+        }
+      } catch (err) {
+        if (this.options.host?.abortOnError === false) {
+          fileLog(
+            `Unable to copy remote types, ${err}`,
+            'extractRemoteTypes',
+            'error',
+          );
+        } else {
+          throw err;
+        }
+      }
     }
   }
 
+  // it must execute after consumeTypes
   async generateTypes() {
     try {
       const { options } = this;
@@ -132,6 +146,21 @@ class DTSManager {
 
       if (!Object.keys(mapComponentsToExpose).length) {
         return;
+      }
+
+      if (tsConfig.compilerOptions.tsBuildInfoFile) {
+        try {
+          const tsBuildInfoFile = path.resolve(
+            remoteOptions.context,
+            tsConfig.compilerOptions.tsBuildInfoFile,
+          );
+          const mfTypesPath = retrieveMfTypesPath(tsConfig, remoteOptions);
+          if (!fs.existsSync(mfTypesPath)) {
+            fs.rmSync(tsBuildInfoFile, { force: true });
+          }
+        } catch (e) {
+          //noop
+        }
       }
 
       await this.extractRemoteTypes({
@@ -150,7 +179,6 @@ class DTSManager {
         apiTypesPath = retrieveMfAPITypesPath(tsConfig, remoteOptions);
         fs.writeFileSync(apiTypesPath, apiTypes);
       }
-
       try {
         if (remoteOptions.deleteTypesFolder) {
           await rm(retrieveMfTypesPath(tsConfig, remoteOptions), {
@@ -167,7 +195,7 @@ class DTSManager {
     } catch (error) {
       if (this.options.remote?.abortOnError === false) {
         if (this.options.displayErrorInTerminal) {
-          logger.error(`Unable to compile federated types${error}`);
+          logger.error(`Unable to compile federated types ${error}`);
         }
       } else {
         throw error;
@@ -177,13 +205,17 @@ class DTSManager {
 
   async requestRemoteManifest(
     remoteInfo: RemoteInfo,
+    hostOptions: Required<HostOptions>,
   ): Promise<Required<RemoteInfo>> {
     try {
       if (!remoteInfo.url.includes(MANIFEST_EXT)) {
         return remoteInfo as Required<RemoteInfo>;
       }
+      if (remoteInfo.zipUrl) {
+        return remoteInfo as Required<RemoteInfo>;
+      }
       const url = remoteInfo.url;
-      const res = await axiosGet(url);
+      const res = await axiosGet(url, { timeout: hostOptions.timeout });
       const manifestJson = res.data as unknown as Manifest;
       if (!manifestJson.metaData.types.zip) {
         throw new Error(`Can not get ${remoteInfo.name}'s types archive url!`);
@@ -249,6 +281,7 @@ class DTSManager {
   async downloadAPITypes(
     remoteInfo: Required<RemoteInfo>,
     destinationPath: string,
+    hostOptions: Required<HostOptions>,
   ) {
     const { apiTypeUrl } = remoteInfo;
     if (!apiTypeUrl) {
@@ -256,7 +289,7 @@ class DTSManager {
     }
     try {
       const url = apiTypeUrl;
-      const res = await axiosGet(url);
+      const res = await axiosGet(url, { timeout: hostOptions.timeout });
       let apiTypeFile = res.data as string;
       apiTypeFile = apiTypeFile.replaceAll(
         REMOTE_ALIAS_IDENTIFIER,
@@ -358,8 +391,10 @@ class DTSManager {
       async (item) => {
         const remoteInfo = item[1];
         if (!this.remoteAliasMap[remoteInfo.alias]) {
-          const requiredRemoteInfo =
-            await this.requestRemoteManifest(remoteInfo);
+          const requiredRemoteInfo = await this.requestRemoteManifest(
+            remoteInfo,
+            hostOptions,
+          );
           this.remoteAliasMap[remoteInfo.alias] = requiredRemoteInfo;
         }
 
@@ -403,7 +438,11 @@ class DTSManager {
               return;
             }
 
-            await this.downloadAPITypes(remoteInfo, destinationPath);
+            await this.downloadAPITypes(
+              remoteInfo,
+              destinationPath,
+              hostOptions,
+            );
           }),
         );
         this.consumeAPITypes(hostOptions);
@@ -476,6 +515,7 @@ class DTSManager {
           const addNew = await this.downloadAPITypes(
             requiredRemoteInfo,
             destinationPath,
+            hostOptions,
           );
           if (addNew) {
             this.consumeAPITypes(hostOptions);
@@ -500,8 +540,10 @@ class DTSManager {
           );
           if (remoteInfo) {
             if (!this.remoteAliasMap[remoteInfo.alias]) {
-              const requiredRemoteInfo =
-                await this.requestRemoteManifest(remoteInfo);
+              const requiredRemoteInfo = await this.requestRemoteManifest(
+                remoteInfo,
+                hostOptions,
+              );
               this.remoteAliasMap[remoteInfo.alias] = requiredRemoteInfo;
             }
             await consumeTypes(this.remoteAliasMap[remoteInfo.alias]);
@@ -519,7 +561,7 @@ class DTSManager {
               });
               fileLog(`start request manifest`, 'consumeTypes', 'info');
               this.updatedRemoteInfos[updatedRemoteInfo.name] =
-                await this.requestRemoteManifest(parsedRemoteInfo);
+                await this.requestRemoteManifest(parsedRemoteInfo, hostOptions);
               fileLog(
                 `end request manifest, this.updatedRemoteInfos[updatedRemoteInfo.name]: ${JSON.stringify(
                   this.updatedRemoteInfos[updatedRemoteInfo.name],